refactor: reorganize main application flow and extract auth layer into separate module
